+++ src/RcppExports.cpp
@@ -72,15 +72,18 @@ BEGIN_RCPP
 END_RCPP
 }
 // CPP_resample
-Rcpp::S4 CPP_resample(Rcpp::S4& rast_in, Rcpp::S4& rast_out, const Rcpp::StringVector& stat);
-RcppExport SEXP _exactextractr_CPP_resample(SEXP rast_inSEXP, SEXP rast_outSEXP, SEXP statSEXP) {
+Rcpp::S4 CPP_resample(Rcpp::S4& rast_in, Rcpp::S4& rast_out, Rcpp::Nullable<Rcpp::CharacterVector> p_stat, Rcpp::Nullable<Rcpp::Function> p_fun, bool coverage_area, std::string area_method);
+RcppExport SEXP _exactextractr_CPP_resample(SEXP rast_inSEXP, SEXP rast_outSEXP, SEXP p_statSEXP, SEXP p_funSEXP, SEXP coverage_areaSEXP, SEXP area_methodSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< Rcpp::S4& >::type rast_in(rast_inSEXP);
     Rcpp::traits::input_parameter< Rcpp::S4& >::type rast_out(rast_outSEXP);
-    Rcpp::traits::input_parameter< const Rcpp::StringVector& >::type stat(statSEXP);
-    rcpp_result_gen = Rcpp::wrap(CPP_resample(rast_in, rast_out, stat));
+    Rcpp::traits::input_parameter< Rcpp::Nullable<Rcpp::CharacterVector> >::type p_stat(p_statSEXP);
+    Rcpp::traits::input_parameter< Rcpp::Nullable<Rcpp::Function> >::type p_fun(p_funSEXP);
+    Rcpp::traits::input_parameter< bool >::type coverage_area(coverage_areaSEXP);
+    Rcpp::traits::input_parameter< std::string >::type area_method(area_methodSEXP);
+    rcpp_result_gen = Rcpp::wrap(CPP_resample(rast_in, rast_out, p_stat, p_fun, coverage_area, area_method));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -89,7 +92,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_exactextractr_CPP_coverage_fraction", (DL_FUNC) &_exactextractr_CPP_coverage_fraction, 3},
     {"_exactextractr_CPP_exact_extract", (DL_FUNC) &_exactextractr_CPP_exact_extract, 17},
     {"_exactextractr_CPP_stats", (DL_FUNC) &_exactextractr_CPP_stats, 11},
-    {"_exactextractr_CPP_resample", (DL_FUNC) &_exactextractr_CPP_resample, 3},
+    {"_exactextractr_CPP_resample", (DL_FUNC) &_exactextractr_CPP_resample, 6},
     {NULL, NULL, 0}
 };