+++ src/resample.cpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2018-2021 ISciences, LLC.
+// Copyright (c) 2018-2022 ISciences, LLC.
 // All rights reserved.
 //
 // This software is licensed under the Apache License, Version 2.0 (the "License").
@@ -21,6 +21,7 @@
 
 using exactextract::Box;
 using exactextract::RasterStats;
+using exactextract::RasterView;
 
 // TODO merge with nearly-identical code in exact_extract.cpp
 static double get_stat_value(const RasterStats<double> & stats, const std::string & stat_name) {
@@ -50,12 +51,18 @@ static double get_stat_value(const RasterStats<double> & stats, const std::strin
 // [[Rcpp::export]]
 Rcpp::S4 CPP_resample(Rcpp::S4 & rast_in,
                       Rcpp::S4 & rast_out,
-                      const Rcpp::StringVector & stat) {
+                      Rcpp::Nullable<Rcpp::CharacterVector> p_stat,
+                      Rcpp::Nullable<Rcpp::Function> p_fun,
+                      bool coverage_area,
+                      std::string area_method) {
   try {
     Rcpp::Environment raster = Rcpp::Environment::namespace_env("raster");
     Rcpp::Environment xx = Rcpp::Environment::namespace_env("exactextractr");
     Rcpp::Function rasterFn = raster["raster"];
     Rcpp::Function valuesFn = xx[".setValues"];
+    Rcpp::Function nlyrFn = xx[".numLayers"];
+
+    int numLayers = Rcpp::as<int>(nlyrFn(rast_in));
 
     S4RasterSource rsrc(rast_in);
 
@@ -64,10 +71,21 @@ Rcpp::S4 CPP_resample(Rcpp::S4 & rast_in,
     auto grid_in = make_grid(rast_in);
     auto grid_out = make_grid(rast_out);
 
-    std::string stat_name = Rcpp::as<std::string>(stat[0]);
-    bool store_values = requires_stored_values(stat_name);
+    std::string stat_name;
+    bool store_values = false;
+    bool r_summary_function = false;
+
+    if (p_stat.isNotNull()) {
+      Rcpp::CharacterVector stat = p_stat.get();
+      stat_name = Rcpp::as<std::string>(stat[0]);
+      store_values = requires_stored_values(stat_name);
+    } else {
+      r_summary_function = true;
+    }
 
     Rcpp::NumericMatrix values_out = Rcpp::no_init(grid_out.rows(), grid_out.cols());
+
+    std::vector<std::unique_ptr<NumericVectorRaster>> values(numLayers);
 
     for (size_t row = 0; row < grid_out.rows(); row++) {
       // Read enough source raster data to process an entire destination row at
@@ -77,21 +95,63 @@ Rcpp::S4 CPP_resample(Rcpp::S4 & rast_in,
       auto ymax = y + grid_out.dy();
 
       Box row_box{ grid_out.xmin(), ymin, grid_out.xmax(), ymax };
-      auto values = rsrc.read_box(row_box, 0);
+
+      for (int i = 0; i < numLayers; i++) {
+        values[i] = std::unique_ptr<NumericVectorRaster>(
+          static_cast<NumericVectorRaster*>(rsrc.read_box(row_box, i).release()));
+      }
 
       for (size_t col = 0; col < grid_out.cols(); col++) {
-        RasterStats<double> stats{store_values};
-
         Box cell = grid_cell(grid_out, row, col);
         auto coverage_fraction = raster_cell_intersection(grid_in, cell);
 
         auto& cov_grid = coverage_fraction.grid();
-
-        if (!cov_grid.empty()) {
-          stats.process(coverage_fraction, *values);
+        if (coverage_area) {
+          auto areas = get_area_raster(area_method, cov_grid);
+          for (size_t i = 0; i < coverage_fraction.rows(); i++) {
+            for (size_t j = 0; j < coverage_fraction.cols(); j++) {
+              coverage_fraction(i, j) = coverage_fraction(i, j) * (*areas)(i, j);
+            }
+          }
         }
 
-        values_out(row, col) = get_stat_value(stats, stat_name);
+        if (r_summary_function) {
+          Rcpp::Function summary_fun = p_fun.get();
+
+          // Transform values to same grid as coverage fractions
+          auto coverage_vec = as_vector(coverage_fraction);
+
+          Rcpp::NumericVector result;
+
+          if (numLayers == 1) {
+            RasterView<double> rt(*(values[0]), cov_grid);
+            auto value_vec = as_vector(rt);
+            result = summary_fun(value_vec, coverage_vec);
+          } else {
+            Rcpp::NumericMatrix value_mat = Rcpp::no_init(coverage_vec.size(), numLayers);
+            for (int i = 0; i < numLayers; i++) {
+              RasterView<double> rt(*(values[i]), cov_grid);
+              auto value_vec = as_vector(rt);
+              value_mat(Rcpp::_, i) = value_vec;
+            }
+
+            result = summary_fun(value_mat, coverage_vec);
+          }
+
+          if (result.size() != 1) {
+            Rcpp::stop("Summary function must return a single value");
+          }
+
+          values_out(row, col) = result[0];
+        } else {
+          RasterStats<double> stats{store_values};
+
+          if (!cov_grid.empty()) {
+            stats.process(coverage_fraction, *(values[0]));
+          }
+
+          values_out(row, col) = get_stat_value(stats, stat_name);
+        }
       }
     }